requirements/requirements.dev.txt
# Dev Dependencies
setuptools
ipython
ipdb
pudb
pytest-watch

# Code Quality
flake8
pyproject-flake8
black
isort

# docs
mkdocs

# build
wheel
twine

# typing
mypy
types-setuptools

# Install project as editable
-e .